[#50] Ref: Coordinator 를 파일로 분리, 클래스 이름 변경
